+++ crates/swc_ecma_minifier/tests/terser/compress/collapse_vars/reduce_vars_assign/output.js
@@ -1,4 +1,4 @@
-!(function () {
+!function() {
     var a = 1;
-    (a = [].length), console.log(a);
-})();
+    a = 0, console.log(a);
+}();